rg_remake/server/requirements.txt
openenv-core[core]>=0.2.3
fastapi>=0.115.0
uvicorn>=0.24.0
reasoning_gym>=0.1.25


